<fix>[sdnController]: fix hardware l2 attach to multiple cluster
Resolves: ZSTAC-77733
