+++ sdk/src/main/java/org/zstack/sdk/huawei/imaster/CreateHuaweiIMasterVRouterAction.java
@@ -1,10 +1,10 @@
-package org.zstack.sdk;
+package org.zstack.sdk.huawei.imaster;
 
 import java.util.HashMap;
 import java.util.Map;
 import org.zstack.sdk.*;
 
-public class DeleteHuaweiIMasterVpcAction extends AbstractAction {
+public class CreateHuaweiIMasterVRouterAction extends AbstractAction {
 
     private static final HashMap<String, Parameter> parameterMap = new HashMap<>();
 
@@ -12,7 +12,7 @@ public class DeleteHuaweiIMasterVpcAction extends AbstractAction {
 
     public static class Result {
         public ErrorCode error;
-        public org.zstack.sdk.DeleteHuaweiIMasterVpcResult value;
+        public org.zstack.sdk.huawei.imaster.CreateHuaweiIMasterVRouterResult value;
 
         public Result throwExceptionIfError() {
             if (error != null) {
@@ -25,14 +25,20 @@ public class DeleteHuaweiIMasterVpcAction extends AbstractAction {
         }
     }
 
+    @Param(required = true, maxLength = 255, nonempty = false, nullElements = false, emptyString = true, noTrim = false)
+    public java.lang.String name;
+
+    @Param(required = false, maxLength = 2048, nonempty = false, nullElements = false, emptyString = true, noTrim = false)
+    public java.lang.String description;
+
     @Param(required = true, nonempty = false, nullElements = false, emptyString = true, noTrim = false)
-    public java.lang.String uuid;
+    public java.lang.String huaweiVpcUuid;
 
     @Param(required = false)
-    public java.lang.String sdnControllerUuid;
+    public java.lang.String resourceUuid;
 
-    @Param(required = false)
-    public java.lang.String deleteMode = "Permissive";
+    @Param(required = false, nonempty = false, nullElements = false, emptyString = true, noTrim = false)
+    public java.util.List tagUuids;
 
     @Param(required = false)
     public java.util.List systemTags;
@@ -66,8 +72,8 @@ public class DeleteHuaweiIMasterVpcAction extends AbstractAction {
             return ret;
         }
         
-        org.zstack.sdk.DeleteHuaweiIMasterVpcResult value = res.getResult(org.zstack.sdk.DeleteHuaweiIMasterVpcResult.class);
-        ret.value = value == null ? new org.zstack.sdk.DeleteHuaweiIMasterVpcResult() : value; 
+        org.zstack.sdk.huawei.imaster.CreateHuaweiIMasterVRouterResult value = res.getResult(org.zstack.sdk.huawei.imaster.CreateHuaweiIMasterVRouterResult.class);
+        ret.value = value == null ? new org.zstack.sdk.huawei.imaster.CreateHuaweiIMasterVRouterResult() : value; 
 
         return ret;
     }
@@ -96,11 +102,11 @@ public class DeleteHuaweiIMasterVpcAction extends AbstractAction {
 
     protected RestInfo getRestInfo() {
         RestInfo info = new RestInfo();
-        info.httpMethod = "DELETE";
-        info.path = "/sdn-controller/huawei-imaster/vpcs/{uuid}";
+        info.httpMethod = "POST";
+        info.path = "/sdn-controller/huawei-imaster/vrouters";
         info.needSession = true;
         info.needPoll = true;
-        info.parameterName = "";
+        info.parameterName = "params";
         return info;
     }
