+++ sdk/src/main/java/org/zstack/sdk/huawei/imaster/DeleteHuaweiIMasterVpcResult.java
@@ -1,4 +1,4 @@
-package org.zstack.sdk;
+package org.zstack.sdk.huawei.imaster;
