+++ sdk/src/main/java/org/zstack/sdk/huawei/imaster/HuaweiIMasterFabricInventory.java
@@ -1,4 +1,4 @@
-package org.zstack.sdk;
+package org.zstack.sdk.huawei.imaster;
