+++ sdk/src/main/java/org/zstack/sdk/huawei/imaster/QueryHuaweiIMasterVpcResult.java
@@ -1,8 +1,8 @@
-package org.zstack.sdk;
+package org.zstack.sdk.huawei.imaster;
 
 
 
-public class QueryHuaweiIMasterVRouterResult {
+public class QueryHuaweiIMasterVpcResult {
     public java.util.List inventories;
     public void setInventories(java.util.List inventories) {
         this.inventories = inventories;
